refactor(rengine): move dyn compatible traits in visitors
